Core restructuring: Move axis settings to the plotting widgets
Before this, the code related to the axes settings was in the UI class. However, this is becoming a mesh after adding the Smitch chart plot. It is needed to move the axis control to the widgets themselves and so that the main program is in charge simply of the trace management between widgets