fix: Order schematic healing actions by priority
This commit adds:
 - priority to healing actions
 - sorting of actions by priority before returning them to caller

Without sorting it possible to have a sequence of actions which
leaves schematic in incosistent state, for example:
 1. replace node of a port located at (x,y)
 2. move node located at (x,y) to (x',y')

if this action sequence was executed the port would be connected
to a node at different location than the port's, which is wrong.

To avoid this, each action is assigned a priority and all of them
are sorted according to it, so the final sequence ("healing plan")
doesn't leave schematic in broken state.

Fixes ra3xdh#1302