Avoid double-layering wires in certain conditions
The algorithm for providing a node at location is designed in a way
that if a newly created node lies above one or more wires, only one
of the wires is split by the node. In some conditions this can lead
to double-layering of wires.

This commit removes "early exit" after the first splitting of a wire
and makes all wires be considered for splitting.